requirements.txt
# IQ Option API Python Wrapper
# Version: 7.1.3

# Core dependencies
requests>=2.28.0
# Note: Se você tiver o pacote 'iqoptionapi' antigo instalado do PyPI, pode haver conflito
# Solução: pip uninstall iqoptionapi (para usar apenas o código-fonte local)
websocket-client>=1.4.0

# Optional but recommended
python-dateutil>=2.8.0
python-dotenv>=1.0.0

# Web interface
Flask>=3.0.0

# Production server
gunicorn>=21.0.0

# Development dependencies (optional)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=5.0.0
# pylint>=2.15.0

